+++ src/Console/Command/Magento/TestTemplates.php
@@ -69,10 +69,12 @@ class TestTemplates extends AbstractTestCommand
         '2.4.4-p3',
         '2.4.5',
         '2.4.5-p2',
+        '2.4.5-p3',
         '2.4.6',
         '2.4.6-p1',
         '2.4.7-beta1',
-        '2.4.7-beta2'
+        '2.4.7-beta2',
+        '2.4.7-beta3'
     ];
 
     /**