+++ src/Docker/Compose/Composition.php
@@ -351,9 +351,15 @@ class Composition
             . $firstContainerWithName . DIRECTORY_SEPARATOR;
         $this->prepareDirectoryToDumpComposition($output, $dockerComposeDir, $force);
 
-        $dockerComposeVersion = $compositionYaml[0]['version'];
+        $dockerComposeVersion = $compositionYaml[0]['version'] ?? '3.7';
         $compositionYaml = ArrayHelper::arrayMergeReplaceRecursive(...$compositionYaml);
-        $compositionYaml['version'] = $dockerComposeVersion;
+
+        // `version` is obsolete in Docker Compose V2
+        if ($this->dockerCompose->isDockerComposeV2()) {
+            unset($compositionYaml['version']);
+        } else {
+            $compositionYaml['version'] = $dockerComposeVersion;
+        }
 
         // Parse, compile and combine all dev tools into one array
         $devToolsYaml = array_map(
@@ -365,8 +371,13 @@ class Composition
             $devToolsYaml = ArrayHelper::arrayMergeReplaceRecursive(...$devToolsYaml);
         }
 
+        // `version` is obsolete in Docker Compose V2
         if (!empty($devToolsYaml)) {
-            $devToolsYaml['version'] = $dockerComposeVersion;
+            if ($this->dockerCompose->isDockerComposeV2()) {
+                unset($devToolsYaml['version']);
+            } else {
+                $devToolsYaml['version'] = $dockerComposeVersion;
+            }
         }
 
         $modificationContext = $this->prepareContext(